+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for the [`ghc-typelits-knownnat`](http://hackage.haskell.org/package/ghc-typelits-knownnat) package
 
+## 0.2.3
+* Solve normalised literal constraints, i.e.:
+  * `KnownNat (((addrSize + 1) - (addrSize - 1))) ~ KnownNat 2`
+
 ## 0.2.2 *September 29th 2016*
 * New features:
   * Derive smaller constraints from larger constraints when they differ by a single variable, i.e. `KnownNat (a + b), KnownNat b` implies `KnownNat a`.